# Speedup embedding plots when len(cluster_colors) < len(cluster_categories)